--- a/pkg/bufferdecoder/protocol_test.go
+++ b/pkg/bufferdecoder/protocol_test.go
@@ -15,35 +15,47 @@ import (
 // If so, then this means you need to update the decoder functions and GetSizeBytes functions.
 
 func TestContextSize(t *testing.T) {
+	t.Parallel()
+
 	var v Context
 	size := int(unsafe.Sizeof(v))
 	assert.Equal(t, size, int(v.GetSizeBytes()))
 }
 func TestChunkMetaSize(t *testing.T) {
+	t.Parallel()
+
 	var v ChunkMeta
 	size := int(unsafe.Sizeof(v))
 	assert.Equal(t, size-7, int(v.GetSizeBytes()))
 }
 
 func TestVfsWriteMetaSize(t *testing.T) {
+	t.Parallel()
+
 	var v VfsFileMeta
 	size := int(unsafe.Sizeof(v))
 	assert.Equal(t, size-4, int(v.GetSizeBytes()))
 }
 
 func TestKernelModuleMetaSize(t *testing.T) {
+	t.Parallel()
+
 	var v KernelModuleMeta
 	size := int(unsafe.Sizeof(v))
 	assert.Equal(t, size-4, int(v.GetSizeBytes()))
 }
 
 func TestBpfObjectMetaSize(t *testing.T) {
+	t.Parallel()
+
 	var v BpfObjectMeta
 	size := int(unsafe.Sizeof(v))
 	assert.Equal(t, size, int(v.GetSizeBytes()))
 }
 
 func TestMprotectWriteMetaSize(t *testing.T) {
+	t.Parallel()
+
 	var v MprotectWriteMeta
 	size := int(unsafe.Sizeof(v))
 	assert.Equal(t, size-4, int(v.GetSizeBytes()))