--- a/pkg/changelog/changelog_test.go
+++ b/pkg/changelog/changelog_test.go
@@ -7,6 +7,8 @@ import (
 )
 
 func TestChangelog(t *testing.T) {
+	t.Parallel()
+
 	cl := NewChangelog[int]()
 
 	// Test GetCurrent on an empty changelog