--- a/pkg/cmd/flags/tracee_ebpf_output_test.go
+++ b/pkg/cmd/flags/tracee_ebpf_output_test.go
@@ -10,6 +10,8 @@ import (
 )
 
 func TestPrepareTraceeEbpfOutput(t *testing.T) {
+	t.Parallel()
+
 	testCases := []struct {
 		testName       string
 		outputSlice    []string
@@ -149,7 +151,11 @@ func TestPrepareTraceeEbpfOutput(t *testing.T) {
 		},
 	}
 	for _, testcase := range testCases {
+		testcase := testcase
+
 		t.Run(testcase.testName, func(t *testing.T) {
+			t.Parallel()
+
 			output, err := TraceeEbpfPrepareOutput(testcase.outputSlice, false)
 			if err != nil {
 				assert.ErrorContains(t, err, testcase.expectedError.Error())