--- a/pkg/cmd/initialize/sigs_test.go
+++ b/pkg/cmd/initialize/sigs_test.go
@@ -12,12 +12,16 @@ import (
 )
 
 func Test_CreateEventsFromSigs(t *testing.T) {
+	t.Parallel()
+
 	tests := []struct {
+		name       string
 		startId    events.ID
 		signatures []detect.Signature
 		expected   []events.Definition
 	}{
 		{
+			name:    "fake_event_0",
 			startId: events.ID(6001),
 			signatures: []detect.Signature{
 				newFakeSignature(
@@ -50,6 +54,7 @@ func Test_CreateEventsFromSigs(t *testing.T) {
 			},
 		},
 		{
+			name:    "fake_event_1/2",
 			startId: events.ID(6010),
 			signatures: []detect.Signature{
 				newFakeSignature(
@@ -106,6 +111,7 @@ func Test_CreateEventsFromSigs(t *testing.T) {
 			},
 		},
 		{
+			name:    "fake_event_3",
 			startId: events.ID(6100),
 			signatures: []detect.Signature{
 				newFakeSignature(
@@ -144,26 +150,32 @@ func Test_CreateEventsFromSigs(t *testing.T) {
 	}
 
 	for _, test := range tests {
-		CreateEventsFromSignatures(test.startId, test.signatures)
+		test := test
 
-		for _, expected := range test.expected {
-			eventDefID, ok := events.Core.GetDefinitionIDByName(expected.GetName())
-			assert.True(t, ok)
-			eventDefinition := events.Core.GetDefinitionByID(eventDefID)
+		t.Run(test.name, func(t *testing.T) {
+			t.Parallel()
 
-			assert.Equal(t, expected.GetID(), eventDefinition.GetID())
-			assert.Equal(t, expected.GetID32Bit(), eventDefinition.GetID32Bit())
-			assert.Equal(t, expected.GetName(), eventDefinition.GetName())
-			assert.Equal(t, expected.GetDocPath(), eventDefinition.GetDocPath())
-			assert.Equal(t, expected.IsInternal(), eventDefinition.IsInternal())
-			assert.Equal(t, expected.IsSyscall(), eventDefinition.IsSyscall())
-			assert.ElementsMatch(t, expected.GetSets(), eventDefinition.GetSets())
-			assert.ElementsMatch(t, expected.GetParams(), eventDefinition.GetParams())
+			CreateEventsFromSignatures(test.startId, test.signatures)
 
-			dependencies := eventDefinition.GetDependencies()
-			expDependencies := expected.GetDependencies()
-			assert.ElementsMatch(t, expDependencies.GetIDs(), dependencies.GetIDs())
-		}
+			for _, expected := range test.expected {
+				eventDefID, ok := events.Core.GetDefinitionIDByName(expected.GetName())
+				assert.True(t, ok)
+				eventDefinition := events.Core.GetDefinitionByID(eventDefID)
+
+				assert.Equal(t, expected.GetID(), eventDefinition.GetID())
+				assert.Equal(t, expected.GetID32Bit(), eventDefinition.GetID32Bit())
+				assert.Equal(t, expected.GetName(), eventDefinition.GetName())
+				assert.Equal(t, expected.GetDocPath(), eventDefinition.GetDocPath())
+				assert.Equal(t, expected.IsInternal(), eventDefinition.IsInternal())
+				assert.Equal(t, expected.IsSyscall(), eventDefinition.IsSyscall())
+				assert.ElementsMatch(t, expected.GetSets(), eventDefinition.GetSets())
+				assert.ElementsMatch(t, expected.GetParams(), eventDefinition.GetParams())
+
+				dependencies := eventDefinition.GetDependencies()
+				expDependencies := expected.GetDependencies()
+				assert.ElementsMatch(t, expDependencies.GetIDs(), dependencies.GetIDs())
+			}
+		})
 	}
 }