@rafaeldtinoco 2023-10-03
> Currently, the use of t.Parallel() doesn't affect the performance of the tests (slight difference):

Where ? In runners or your local environment ? Because if in a SMP environment we aren't benefiting from having it set, we shouldn't set it IMO.