@geyslan 2023-10-03
> > Currently, the use of t.Parallel() doesn't affect the performance of the tests (slight difference):
> 
> Where ? In runners or your local environment ? Because if in a SMP environment we aren't benefiting from having it set, we shouldn't set it IMO.

Local tests shown 1s of improvement, in average, that's why I put that it didn't affect the performance. On Github workflow is slightly similar. But as I pointed in the commit:

```
Critically, this practice elevates the detection of insidious race
conditions and state leakages—issues that often remain concealed during
serial test execution.
```

The main reason for using `t.Parallel()` is to have more checks (i.e.: DATA RACE) when building unit tests.