@rafaeldtinoco 2023-10-03
> The main reason for using `t.Parallel()` is to have more checks (i.e.: DATA RACE) when building unit tests.

Fair enough.